Updated automations to use 64-bit IDs, and now returning the ID after adding a new automation

+++ common/src/lib.rs
@@ -3,9 +3,9 @@ use zvariant::Type;
 
 /* === Triggers === */
 
-pub const TRIGGER_TIME: u32 = 1;
+pub const TRIGGER_TIME: i64 = 1;
 
-pub fn trigger_to_name(x: u32) -> String {
+pub fn trigger_to_name(x: i64) -> String {
     match x {
         TRIGGER_TIME => "Time-based".to_string(),
         _ => "Unknown".to_string()
@@ -14,16 +14,16 @@ pub fn trigger_to_name(x: u32) -> String {
 
 #[derive(Serialize, Type)]
 pub struct AutomationTrigger {
-    pub id: u32,
+    pub id: i64,
     pub json: String,
 }
 
 /* === Actions === */
 
-pub const ACTION_COMMAND: u32 = 1;
-pub const ACTION_NOTIFICATION: u32 = 2;
+pub const ACTION_COMMAND: i64 = 1;
+pub const ACTION_NOTIFICATION: i64 = 2;
 
-pub fn action_to_name(x: u32) -> String {
+pub fn action_to_name(x: i64) -> String {
     match x {
         ACTION_COMMAND => "Run command".to_string(),
         ACTION_NOTIFICATION => "Send notification".to_string(),
@@ -33,7 +33,7 @@ pub fn action_to_name(x: u32) -> String {
 
 #[derive(Serialize, Type)]
 pub struct AutomationAction {
-    pub id: u32,
+    pub id: i64,
     pub json: String,
 }
